include clientes module, with screens, services and routes
