+++ src/app/modules/clientes/clientes-routing.module.ts
@@ -1,0 +1,26 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { ClientesComponent } from './clientes/clientes.component';
+import { ClienteDetalleComponent } from './cliente-detalle/cliente-detalle.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    redirectTo: '/clientes',
+    pathMatch: 'full'
+  },
+  {
+    path: 'clientes',
+    component: ClientesComponent
+  },
+  {
+    path: 'clientes/:id',
+    component: ClienteDetalleComponent
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class ClientesRoutingModule { }
