+++ src/app/modules/clientes/clientes/clientes.component.spec.ts
@@ -1,0 +1,24 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ClientesComponent } from './clientes.component';
+import { ClienteDetalleComponent } from '../cliente-detalle/cliente-detalle.component';
+
+describe('ClientesComponent', () => {
+  let component: ClientesComponent;
+  let fixture: ComponentFixture<ClientesComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [ClientesComponent]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(ClientesComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
