+++ src/app/modules/clientes/services/cliente-service.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ClienteServiceService } from './cliente-service.service';
+
+describe('ClienteServiceService', () => {
+  let service: ClienteServiceService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ClienteServiceService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
